% Generated by roxygen2 (4.0.1): do not edit by hand
\name{bscore_calc}
\alias{bscore_calc}
\title{Compute psi_xp(B_i, X_i, theta)}
\usage{
bscore_calc(predictors, B, groups, theta, data, set.NA.to.0 = TRUE)
}
\arguments{
\item{predictors}{character vector of predictors from model}

\item{B}{character naming B variable in data}

\item{G}{character vector (length == 1) naming group variable in data}

\item{theta}{}

\item{data}{dataframe}
}
\value{
N X length(theta) matrix of scores
}
\description{
Compute psi_xp(B_i, X_i, theta)
}

