% Generated by roxygen2 (4.0.1): do not edit by hand
\name{calc_effect}
\alias{calc_effect}
\title{Calculate Effect estimates}
\usage{
calc_effect(obj, alpha1, trt.lvl1, alpha2, trt.lvl2, effect, marginal,
  conf.level = 0.95, print = FALSE)
}
\arguments{
\item{TBD}{}
}
\value{
TBD
}
\description{
Calculate Effect estimates
}

