% Generated by roxygen2 (4.0.1): do not edit by hand
\name{calc_estimands}
\alias{calc_estimands}
\title{Calculate estimands}
\usage{
calc_estimands(potential_outcomes, alphas)
}
\arguments{
\item{potential_outcomes}{a list of potential outcomes generated by
\code{\link{generate_po}}}

\item{alphas}{a vector of 'allocation schemes' in (0, 1)}
}
\value{
a data frame with 4 columns: 1) alpha (the allocation scheme) 2)
a0 - the outcome estimand for untreated, 3) a1 - the outcome estimand for
treated, 4) marg - the marginal outcome estimand
}
\description{
Takes a set of potential outcomes generated by \code{\link{generate_po}} and
returns a data frame containing the population level causal estimands
}
\examples{
calc_estimands(generate_po(sim_base_data(3, 40)), seq(.2, .8, by = .1))
}

