% Generated by roxygen2 (4.0.1): do not edit by hand
\name{compare_truth}
\alias{compare_truth}
\title{Compare simulation estimates to truth}
\usage{
compare_truth(a1, trt1 = NA, a2 = NA, trt2 = NA, effect, marginal,
  list_of_simulation_estimates, truth)
}
\description{
Compare simulation estimates to truth
}

