% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{diagnose_weights}
\alias{diagnose_weights}
\title{Plot histograms of weights from an interference object}
\usage{
diagnose_weights(obj, allocations = NULL, ...)
}
\arguments{
\item{obj}{an \code{interference} object}

\item{allocations}{optional numeric vector of allocations for which to print
histogram. If NULL (the default), five allocations selected evenly from the 
first allocation to the last are printed.}

\item{...}{additional arguments passed to \code{\link{hist}}}
}
\value{
histogram of group-level weights
}
\description{
Plot histograms of weights from an interference object
}
