% Generated by roxygen2 (4.0.1): do not edit by hand
\name{generate_po}
\alias{generate_po}
\title{Generate Potential Outcomes}
\usage{
generate_po(base_data, parameters = c(0.5, -0.788, -2.953, -0.098, -0.145,
  0.351))
}
\arguments{
\item{base_data}{simulated data from \code{\link{sim_base_data}}}

\item{parameters}{vector of parameters used to generate potential outcomes.
Defaults to parameters used in Perez et. al 2014.}
}
\value{
a list with length(# of groups) where each entry is a n_i x n_i x 2
array representing potential outcomes for each subject by # of other subject's
treated by 2 treatment levels {0, 1}
}
\description{
Create a list of arrays representing potential outcomes \eqn{y_{ij}(a, k)},
(where \eqn{a} is the treatment level and \eqn{k} is the number of other
subjects in a group receiving treatment).
}
\details{
Potential outcomes are generated as bernoulli random variables with expectation:
\deqn{logit^{-1}(X\psi)}{plogis(X \%*\% parameters)}. \eqn{X_{ij}} is the vector:
\eqn{(1, a, \alpha, X_{1ij}, X_{2ij}, a * \alpha)}{(1, a, \alpha, X_{1ij},
X_{2ij}, a * \alpha)} where \eqn{\alpha = (a + k)/n_i}.
}

