% Generated by roxygen2 (4.0.1): do not edit by hand
\name{get_args}
\alias{get_args}
\title{Get Arguments}
\usage{
get_args(FUN, ...)
}
\arguments{
\item{FUN}{function for which to find arguments}

\item{...}{any arguments. Those necessary for FUN must be named as appropriate for FUN}
}
\value{
list of arguments for FUN
}
\description{
Extracts the names of the arguments from a function, and creates a list
of those arguments where they exist in ... .
}
\examples{
myargs <- get_args(lm, formula = Sepal.Length ~ Sepal.Width, data = iris )
summary(do.call('lm', myargs))
}

