% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/utilities.R
\name{group_means}
\alias{group_means}
\title{Calculate outcome mean per group per treatment level}
\usage{
group_means(Y, A, G, a = NA)
}
\arguments{
\item{Y}{vector of outcomes}

\item{A}{vector of treatment assignments}

\item{G}{vector of group assignments}

\item{a}{value of treatment level, defaults to NA. NA is used for marginal
estimates.}
}
\value{
matrix of group means
}
\description{
Calculate outcome mean per group per treatment level
}

