% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inferference.R
\docType{package}
\name{inferference}
\alias{inferference}
\title{Methods for causal inference with interference}
\description{
Interference occurs when the treatment of one unit affects outcomes of other
units. This package provides methods for estimating causal effects in the 
presence of interference. Currently it implements the IPW estimators 
proposed by Tchetgen Tchetgen and Vanderweele (2012) (\doi{10.1177/0962280210386779})
and developed further in Heydrich-Perez et al. (2014) (\doi{10.1111/biom.12184}).
}
\references{
Saul, B. and Hugdens, M. G. (2017). 
A Recipe for {inferference}: {S}tart with Causal Inference. {A}dd Interference. {M}ix Well with {R}. 
Journal of Statistical Software, 82(2), 1-21. \doi{10.18637/jss.v082.i02}
}
