% Generated by roxygen2 (4.0.1): do not edit by hand
\name{interference}
\alias{interference}
\title{Run Interference}
\usage{
interference(estimation_type = "ipw", allocations, data, groups, outcome,
  treatment, propensityB = treatment, propensity_formula, conf.level = 0.95,
  rescale.factor = 1, ...)
}
\arguments{
\item{estimation_type}{currently only supports `ipw`.}

\item{allocations}{the allocation (coverage) schemes for which to estimate effects.
Must be in (0, 1].}

\item{data}{the analysis data.frame}

\item{groups}{quoted name of group variable in \code{data}}

\item{outcome}{quoted name of outcome variable in \code{data}}

\item{treatment}{quoted name of treatment variable in \code{data}}

\item{propensityB}{Optional. If the treatment variable is different for
\code{integrand} and \code{likelihood}, include the quote name of the variable
to be passed to \code{likelihood}. Defaults to \code{treatment}.}

\item{propensity_formula}{formula passed to model_method for estimation of propensity score
parameters.}

\item{rescale.factor}{factor by which to rescale values. Defaults to 1.}

\item{conf.level}{Confidence level for confidence intervals. Defaults to 0.95.}

\item{...}{additional arguments passed to other functions such as
\code{\link{glmer}}, \code{\link{grad}}, and \code{integrand} or \code{likelihood}.}
}
\value{
Returns a list of overall and group-level IPW point estimates
(the output of \code{\link{ipw_point_estimates}}), overall and group-level IPW
point estimates (using the weight derivatives), scores (the output of
\code{\link{score_matrix_calc}}), the computed weight matrix, the computed
weight derivative array, and a summary.
}
\description{
Prepares the object necessary to compute IPW effect estimates with
\code{\link{calc_effect}}, \code{\link{direct_effect}}, \code{\link{indirect_effect}},
\code{\link{total_effect}}, or \code{\link{overall_effect}}.
}

