% Generated by roxygen2 (4.0.1): do not edit by hand
\name{interference_sample}
\alias{interference_sample}
\title{Interference Sample Data}
\format{a dataset with 6 variables and 3000 rows
\describe{
   \item{y}{the outcome}
   \item{X1}{a covariate}
   \item{X2}{another covariate}
 }}
\description{
A sample dataset.
}

