% Generated by roxygen2 (4.0.1): do not edit by hand
\name{ipw_estimates}
\alias{ipw_estimates}
\title{Create IPW estimates}
\usage{
ipw_estimates(y, G, A, data, weights, weight_dervs, predictors, rescale.factor,
  set.NA.to.0 = TRUE)
}
\arguments{
\item{y}{vector of unweighted group means (i.e. output from \code{\link{group_means}}}

\item{a}{treatment level (0,1) for which to compute weights. Defaults to NULL which returns marginal.}

\item{weights}{weight matrix to use}

\item{rescale.factor}{factor by which to rescale values}

\item{na.rm}{exclude groups missing weights? defaults to FALSE.}
}
\value{
length(alpha) vector of IPW estimates
}
\description{
Create IPW estimates
}

