% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ipw_interference.R
\name{ipw_interference}
\alias{ipw_interference}
\title{IPW Interference estimation}
\usage{
ipw_interference(propensity_integrand,
  loglihood_integrand = propensity_integrand, allocations, Y, X, A, B = A,
  G, fixed.effects, random.effects, propensity_formula, variance_estimation,
  set_NA_to_0 = TRUE, ...)
}
\arguments{
\item{propensity_integrand}{function used as the integrand in computing the IPW weights.
Defaults to \code{\link{logit_integrand}}.}

\item{loglihood_integrand}{function used as the integrand in computing the scores in
\code{\link{score_calc}}. Defaults to \code{\link{logit_integrand}}.}

\item{allocations}{the allocation (coverage) schemes for which to estimate effects.
Must be in [0, 1].}

\item{Y}{outcome vector}

\item{X}{covariate matrix}

\item{A}{treatmeent vector}

\item{B}{'participation' vector. Defaults to A in the case there is no
participation variable.}

\item{G}{group assignment vector}

\item{fixed.effects}{vector of fixed effects}

\item{random.effects}{vector of random effects}

\item{variance_estimation}{currently supports 'robust' or 'simple'}

\item{set_NA_to_0}{if TRUE, sets any weights that returned an NA value to 0.
Defaults to TRUE}

\item{...}{additional arguments passed to other functions such as
\code{\link{glmer}}, \code{\link{grad}}, and \code{integrand} or \code{likelihood}.}
}
\value{
Returns a list of overall and group-level IPW point estimates
(the output of \code{\link{ipw_point_estimates}}), overall and group-level IPW
point estimates (using the weight derivatives), scores (the output of
\code{\link{score_matrix_calc}}), the computed weight matrix, and the computed
weight derivative array.
}
\description{
Prepares the object necessary to compute IPW effect estimates with
\code{\link{calc_effect}}.
}

