% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ipw_point_estimates.R
\name{ipw_point_estimates}
\alias{ipw_point_estimates}
\title{Calculate IPW point estimates}
\usage{
ipw_point_estimates(Y, G, A, weights)
}
\arguments{
\item{Y}{vector of outcomes}

\item{G}{vector of group assignments}

\item{A}{vector of treatment assignments}

\item{weights}{weight matrix/array to use from either \code{\link{wght_matrix}}
or \code{\link{wght_deriv_array}}}
}
\value{
list containing point estimates for marginal outcomes and estimates
per treatment level
}
\description{
Calculate IPW point estimates
}

