% Generated by roxygen2 (4.0.1): do not edit by hand
\name{log_likelihood}
\alias{log_likelihood}
\title{Log Likelihood}
\usage{
log_likelihood(x, pos, integrand = logit_integrand, ...)
}
\arguments{
\item{x}{}

\item{pos}{}

\item{integrand}{Defaults to logit_integrand}
}
\value{
value of log likelihood
}
\description{
Used by \code{\link{score_matrix_calc}} to compute the log likelihood.
}

