% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/utilities.R
\name{oe}
\alias{oe}
\title{Retrieve Overall Effect Estimates}
\usage{
oe(object, allocation1, allocation2, value = c("estimate", "std.error",
  "conf.low", "conf.high"))
}
\arguments{
\item{object}{an object of class \code{interference}}

\item{allocation1}{the allocation scheme for which to estimate overall effects}

\item{allocation2}{the allocation scheme for which to estimate overall effects}

\item{value}{character vector of values to return. Defaults to
\code{c('estimate', 'std.error', 'conf.low', 'conf.high')}}
}
\value{
a data.frame with a single row with requested values
}
\description{
Retrieves the population average overall causal effect:
\eqn{\hat{Y}(alpha1) - \hat{Y}(lpha2)}{Yhat(alpha1) - Yhat(alpha2)}
}

