% Generated by roxygen2 (4.0.1): do not edit by hand
\name{PrAX_integrand}
\alias{PrAX_integrand}
\title{Calculate Pr(A|X)}
\usage{
PrAX_integrand(b, x, pos = NA, X, A, theta, alpha, r = 2/3, type)
}
\arguments{
\item{b}{random effect necessary for \code{\link{integrate}}}

\item{X}{n by length(theta) - 1 matrix of covariates. Make sure the order of columns in X corresponds to theta}

\item{theta}{p + 1 vector of fixed effects plus the random effect variance. The variance estimate must be last.}

\item{A}{vector of observed treatments (0,1)}

\item{alpha}{The allocation strategy. This can be empty if type = 'c'.}

\item{r}{Randomization probability defaults to 2/3}

\item{type}{Either 'b' for including alpha in the product or 'c' does not include alpha.}

\item{x}{The argument passed to be \code{\link{grad}}. Only used if \code{pos} is not NA.}

\item{pos}{The position of theta for which to take the derivative. Defaults to NA.}
}
\value{
value of the integrand
}
\description{
\code{grod * dnorm(sd = sqrt(variance))}
}

