% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{print.interference}
\alias{print.interference}
\title{Prints a summary of an interference object}
\usage{
\method{print}{interference}(x, ...)
}
\arguments{
\item{x}{object of class 'interference'}

\item{...}{ignored}
}
\description{
Prints a summary of an interference object
}
