% Generated by roxygen2 (4.0.1): do not edit by hand
\name{run_interference}
\alias{run_interference}
\title{Run Interference}
\usage{
run_interference(f.ab, alphas, data, groups, outcome, A, B, predictors,
  type = "b", rescale.factor = 1)
}
\arguments{
\item{predictors}{character vector of predictors from model}

\item{B}{character naming B variable in data}

\item{G}{character vector (length == 1) naming group variable in data}

\item{theta}{}

\item{data}{dataframe}
}
\value{
N X length(theta) matrix of scores
}
\description{
Run Interference
}

