% Generated by roxygen2 (4.0.1): do not edit by hand
\name{score_calc}
\alias{score_calc}
\title{Compute scores for a single group}
\usage{
score_calc(integrand = logit_integrand, hide.errors = TRUE, fixed.effects,
  random.effect, ...)
}
\arguments{
\item{integrand}{function to used for the integrand.
Defaults to \code{\link{logit_integrand}}}

\item{hide.errors}{Hide errors printed from \code{\link{grad}}.
Defaults to true}

\item{params}{See \code{\link{logit_integrand}}.}

\item{...}{additional arguments pass to the integrand function}
}
\value{
length(theta) vector of scores
}
\description{
Used by \code{\link{score_matrix_calc}} to log likelihood derivatives for
a single group.
}

