% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/score_matrix.R
\name{score_matrix}
\alias{score_matrix}
\title{Calculate matrix of log Likelihood derivatives}
\usage{
score_matrix(integrand = logit_integrand, X, A, G, fixed.effects,
  random.effects, ...)
}
\arguments{
\item{integrand}{function passed to \code{\link{log_likelihood}}. Defaults to
\code{\link{logit_integrand}}}

\item{X}{covariate matrix}

\item{A}{vector of treatment assignments}

\item{G}{vector of group assignments}

\item{fixed.effects}{vector of fixed effect parameters}

\item{random.effects}{OPTIONAL vector random effect parameters}

\item{...}{additional arguments passed to \code{integrand} or \code{\link{grad}}.
For example, one can change the \code{method} argument in \code{grad}.}
}
\value{
N X length(params) matrix of scores
}
\description{
Calculate matrix of log Likelihood derivatives
}

