% Generated by roxygen2 (4.0.1): do not edit by hand
\name{sim_base_data}
\alias{sim_base_data}
\title{Simulate a dataset of n individuals into N groups}
\usage{
sim_base_data(N.groups, n.units, X1.distr = "rexp", X1.args = list(rate =
  1/20), ind.dist.distr = "rnorm", ind.dist.args = list(mean = 0, sd =
  0.05), grp.dist.distr = "rlnorm", grp.dist.args = list(meanlog = 0, sdlog
  = 0.75))
}
\arguments{
\item{N.groups}{the number of groups}

\item{n.units}{the number of subjects (total)}

\item{X1.distr}{distribution function of \code{X1} variable. Defaults to
\code{\link{rexp}}}

\item{X1.args}{arguments passed to \code{X1.distr} function. In Perez et al,
X1 represents an individual's age in decades generated from an exponential
distribution with mean 20}

\item{ind.dist.distr}{distribution function of individual distances. Defaults
to \code{\link{rnorm}}}

\item{ind.distr.args}{arguments passed to \code{ind.dist.distr} function}

\item{grp.dist.distr}{distribution function of group level distances. Defaults
to \code{\link{rlnorm}}.}

\item{grp.distr.args}{arguments passed to \code{grp.dist.distr} function}
}
\value{
a data frame
}
\description{
This is currently not very useful except to simulate a dataset like in
Step 0 of Perez et al's simulation from which potential outcomes
will be generated.
}

