% Generated by roxygen2 (4.0.1): do not edit by hand
\name{sim_groups}
\alias{sim_groups}
\title{Simulate a dataset of N groups}
\usage{
sim_groups(N, n, distance.distr = "rlnorm", distr.mean = 0,
  distr.sd = 0.75)
}
\arguments{
\item{N}{the number of groups}

\item{n}{the number of subjects}

\item{distrance.distr}{a character string defining the distribution from
which group level distances are sampled. Defaults to 'rlnorm'.}

\item{distr.mean}{the mean of the chosen distribution. Defaults to 0 as in
Perez-Heydrich 2014.}

\item{distr.sd}{the standard deviation of the chosen distribution. Defaults
to 0.75.}
}
\value{
a data frame with the size and distance for each group
}
\description{
Randomly sllocates n units to N groups and randomly samples a group distance
according to the chosen distribution
}

