% Generated by roxygen2 (4.0.1): do not edit by hand
\name{sim_units}
\alias{sim_units}
\title{Simulate a dataset of n individuals into N groups}
\usage{
sim_units(N, n, X1.distr = "rexp", X1.args = list(rate = 1/20),
  dist.distr = "rnorm", dist.args = list(mean = 0, sd = 0.05),
  grp.dist.distr = "rlnorm", grp.dist.mean = 0, grp.dist.sd = 0.75)
}
\arguments{
\item{N}{the number of groups}

\item{n}{the number of subjects}

\item{X1.args}{In Perez et al, X1 represents an individual's age in
decades generated from an exponential distribution with mean 20}

\item{distr.mean}{the mean of the chosen distribution. Defaults to 0 as in
Perez et al 2014.}

\item{distr.sd}{the standard deviation of the chosen distribution. Defaults
to 0.75.}
}
\value{
a data frame with the size and distance for each group
}
\description{
This is currently not very useful except to simulate a base dataset like in
Step 0 of Perez et al's simulation
}

