% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/utilities.R
\name{te}
\alias{te}
\title{Retrieve Total Effect estimates}
\usage{
te(object, allocation1, allocation2, trt.lvl1 = 0, trt.lvl2 = 1,
  value = c("estimate", "std.error", "conf.low", "conf.high"))
}
\arguments{
\item{object}{an object of class \code{interference}}

\item{allocation1}{the allocation scheme for which to estimate total effects}

\item{allocation2}{the allocation scheme for which to estimate total effects}

\item{trt.lvl1}{Defaults to 0.}

\item{trt.lvl2}{Defaults to 1.}

\item{value}{character vector of values to return. Defaults to
\code{c('estimate', 'std.error', 'conf.low', 'conf.high')}}
}
\value{
a data.frame with a single row with requested values
}
\description{
Retrieves the population average total causal effect for
specified allocations:
\eqn{\hat{Y}(0, alpha1) - \hat{Y}(1, alpha2)}{Yhat(0, alpha1) - Yhat(1, alpha2)}
}

