% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/effect_estimates.R
\name{total_effect}
\alias{total_effect}
\title{Calculate Total Effect Estimates}
\usage{
total_effect(obj, allocation1, allocation2, trt.lvl1 = 0, trt.lvl2 = 1,
  print = FALSE, rescale.factor = 1, conf.level = 0.95)
}
\arguments{
\item{obj}{the name of the object created by \code{\link{ipw_interference}}}

\item{allocation1}{the allocation scheme for which to estimate total effects}

\item{allocation2}{the allocation scheme for which to estimate total effects}

\item{trt.lvl1}{Defaults to 0.}

\item{trt.lvl2}{Defaults to 1.}

\item{print}{TRUE/FALSE. If TRUE, the point estimates and confidence interval
are printed to the console.}

\item{rescale.factor}{factor by which to rescale values. Defaults to 1.}

\item{conf.level}{Confidence level for confidence intervals. Defaults to 0.95.}
}
\value{
See \code{\link{calc_effect}}.
}
\description{
Compute the population average total causal effect:
\eqn{\hat{Y}(0, alpha1) - \hat{Y}(1, alpha2)}{Yhat(0, alpha1) - Yhat(1, alpha2)}
}

