% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{voters}
\alias{voters}
\title{Voting Contagion Experiment Data}
\format{
a dataset with 21 variables and 7722 rows
\itemize{
   \item{family}{household ID}
   \item{denver}{1 = subject in Denver, 0 = Minneapolis}
   \item{treatment}{1 = voting encouragement, 2 = recycling message, 3 = not contacted}
   \item{reached}{1 = subject answered door, 0 = not}
   \item{hsecontact}{1 = household contacted by canvassers, 0 = not}
   \item{voted02p}{1 = voted in '02 primary, 0 = not}
   \item{party}{party affiliation}
   \item{age}{age}
   \item{gender}{gender}    
 }
}
\description{
A dataset of a voting contagion experiment. See Nickerson (2008) for more
details. The variables used in the package vignette are documented here.
}
\references{
Nickerson, D. W. (2008). Is voting contagious? Evidence from two field experiments. 
 American Political Science Review, 102(01), 49-57. \doi{10.1017/S0003055408080039}
}
\keyword{datasets}
