% Generated by roxygen2 (4.0.1): do not edit by hand
\name{wght_calc}
\alias{wght_calc}
\title{Compute IPW weight}
\usage{
wght_calc(type, A, alpha, ...)
}
\arguments{
\item{type}{see description}

\item{A}{vector of treatments}

\item{alpha}{allocation strategy}

\item{...}{other arguments passed to \code{\link{PrAX_integrand}}}
}
\value{
scalar result of the integral
}
\description{
Calculates the IPW for a single group. Used by \code{\link{wght_matrix}} to
create a matrix of weights for each group and allocation scheme.
}
\details{
Type c =
\deqn{\frac{\prod_{j=1}^n \alpha^A_j (1 - \alpha)^(1-
A_j)}{Pr(A|X)}}{prod(alpha^A(1 - alpha)^A) / integrate(PrAX_integrand)}
Type b =
\deqn{\frac{1}{Pr(A|X)}}{1 / integrate(PrAX_integrand)}

Type b incorporates the numerator of type c into the \eqn{Pr(A|X)} integral,
resulting in a slower computation but more accurate results (especially for large groups)
}

