% Generated by roxygen2 (4.0.1): do not edit by hand
\name{wght_deriv_array}
\alias{wght_deriv_array}
\title{Create an array of group weight derivatives}
\usage{
wght_deriv_array(integrand, allocations, data, groups, predictors, treatment,
  params, ...)
}
\arguments{
\item{allocations}{coverage levels in (0, 1), possibly (probably) vector valued}

\item{data}{data frame}

\item{groups}{quoted string for name of variable in data containing group membership}

\item{predictors}{character vector of names of predictor variables in data}

\item{treatment}{character vector of name of treatment variable in data}

\item{params}{p + 1 vector of fixed effects plus the random effect variance.
The variance estimate must be last.}

\item{...}{additional arguments passed to integrand}
}
\value{
a length(unique(group)) X length(params) X length(alphas) array of
group weight derivatives
}
\description{
Uses \code{\link{wght_deriv_calc}} to compute the weight derivatives for each
group per coverage level
}

