% Generated by roxygen2 (4.0.1): do not edit by hand
\name{wght_deriv_calc}
\alias{wght_deriv_calc}
\title{Compute the derivative(s) of a weight}
\usage{
wght_deriv_calc(theta, A, X, type, alpha)
}
\arguments{
\item{theta}{}
}
\value{
vector of derivatives with respect to element of theta
}
\description{
Compute the derivative(s) of a weight
}

