% Generated by roxygen2 (4.0.1): do not edit by hand
\name{wght_matrix}
\alias{wght_matrix}
\title{Create a matrix of group weights}
\usage{
wght_matrix(alphas, data, groups, predictors, A, theta, type)
}
\arguments{
\item{alphas}{coverage levels in (0, 1), possibly (probably) vector valued}

\item{data}{data frame}

\item{groups}{quoted string for name of variable in data containing group membership}

\item{predictors}{character vector of names of predictor variables in data}

\item{A}{character vector of name of treatment variable in data}

\item{theta}{p + 1 vector of fixed effects plus the random effect variance. The variance estimate must be last.}

\item{type}{type of weight to compute. See \code{\link{wght_calc}}}
}
\value{
a length(unique(group)) X length(alphas) matrix of group weights
}
\description{
- creates a length(group) X l matrix of group weights
}

