% Generated by roxygen2 (4.0.1): do not edit by hand
\name{ybar_i_alpha}
\alias{ybar_i_alpha}
\title{Individual level marginal causal estimand Y(alpha)
Takes the n x 2 matrix returned by ybar_ij_aalpha}
\usage{
ybar_i_alpha(input, alpha)
}
\value{
n x 1 vector of ybar_ij(alpha)
}
\description{
Individual level marginal causal estimand Y(alpha)
Takes the n x 2 matrix returned by ybar_ij_aalpha
}

