% Generated by roxygen2 (4.0.1): do not edit by hand
\name{ybar_ij_aalpha}
\alias{ybar_ij_aalpha}
\title{Individual level causal estimand Y(a, alpha)}
\usage{
ybar_ij_aalpha(outcomes, alpha)
}
\value{
n x 2 matrix with values of ybar_ij(a, alpha)
}
\description{
Takes n X n - 1 x 2 array as input
}

