% Generated by roxygen2 (4.0.1): do not edit by hand
\name{direct_effect}
\alias{direct_effect}
\title{Calculate Direct Effect estimates}
\usage{
direct_effect(obj, alpha, trt.lvl1 = 0, trt.lvl2 = 1, print = FALSE,
  conf.level = 0.95)
}
\arguments{
\item{obj}{the name of the object created by \code{\link{run_interference}}}

\item{alpha}{the allocation scheme for which to estimate direct effects}

\item{trt.lvl1}{Defaults to 0.}

\item{trt.lvl2}{Defaults to 1.}

\item{print}{see \code{\link{calc_effect}}}

\item{conf.level}{see \code{\link{calc_effect}}}
}
\value{
See \code{\link{calc_effect}}.
}
\description{
By default, this function computes:
\eqn{\hat{Y}(0, alpha) - \hat{Y}(1, alpha)}{Yhat(0, alpha) - Yhat(1, alpha)}
}

