% Generated by roxygen2 (4.0.1): do not edit by hand
\name{direct_effect}
\alias{direct_effect}
\title{Calculate Direct effect estimates}
\usage{
direct_effect(obj, alpha, trt.lvl1 = "0", trt.lvl2 = "1", print = FALSE,
  conf.level = 0.95)
}
\arguments{
\item{TBD}{}
}
\value{
TBD
}
\description{
Calculate Direct effect estimates
}

