% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utilities.R
\name{direct_effect}
\alias{direct_effect}
\title{Retreive Direct Effect estimates}
\usage{
direct_effect(object, allocation = NULL, trt.lvl1 = 0)
}
\arguments{
\item{object}{an object of class \code{interference}}

\item{allocation}{the allocation scheme for which to estimate direct effects.
If NULL, then returns all direct effects.}

\item{trt.lvl1}{Defaults to 0.}
}
\value{
a data.frame with requested values
}
\description{
Retrieves the population average direct causal effect for a
specified allocation:
\eqn{\hat{Y}(0, alpha) - \hat{Y}(1, alpha)}{Yhat(0, alpha) - Yhat(1, alpha)}.
}

