% Generated by roxygen2 (4.0.1): do not edit by hand
\name{effect_grid}
\alias{effect_grid}
\title{Create a dataset of arguments for effect estimates}
\usage{
effect_grid(allocations, treatments = c(0, 1))
}
\arguments{
\item{allocations}{vector of allocations}

\item{treatments}{vector of treatments. defaults to \code{c(0 ,1)}}
}
\value{
data.frame with arguments necessary for \code{\link{calc_effect}} to
compute all outcome, direct, indirect, total, and overall effect estimates from
an object created from \code{\link{ipw_interference}}
}
\description{
Create a dataset of arguments for effect estimates
}
\examples{
effect_grid(seq(0,1, by = .1), c(0,1))
}

