% Generated by roxygen2 (4.0.1): do not edit by hand
\name{group_means}
\alias{group_means}
\title{Calculate outcome means per group per treatment level}
\usage{
group_means(Y, A, G, a = NA, data)
}
\arguments{
\item{Y}{name (character vector) of observed outcomes in data}

\item{A}{name of treatment variable}

\item{G}{name of grouping variable in data}

\item{a}{value of treatment level, defaults to NA.}
}
\value{
data dataset to use
}
\description{
Calculate outcome means per group per treatment level
}

