% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/utilities.R
\name{group_means}
\alias{group_means}
\title{Calculate outcome means per group per treatment level}
\usage{
group_means(Y, A, G, a = NA, data)
}
\arguments{
\item{Y}{name (character vector) of observed outcomes in data}

\item{A}{name of treatment variable}

\item{G}{name of grouping variable in data}

\item{a}{value of treatment level, defaults to NA.}

\item{data}{data.frame with variables}
}
\value{
data dataset to use
}
\description{
Calculate outcome means per group per treatment level
}

