% Generated by roxygen2 (4.0.1): do not edit by hand
\name{indirect_effect}
\alias{indirect_effect}
\title{Calculate indirect effect estimates}
\usage{
indirect_effect(obj, alpha1, alpha2, trt.lvl = "0", print = FALSE,
  conf.level = 0.95)
}
\arguments{
\item{TBD}{}
}
\value{
TBD
}
\description{
Calculate indirect effect estimates
}

