% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utilities.R
\name{indirect_effect}
\alias{ie}
\alias{indirect_effect}
\title{Retreive Indirect Effect estimates}
\usage{
indirect_effect(object, allocation1, allocation2 = NULL, trt.lvl = 0)

ie(object, allocation1, allocation2 = NULL, trt.lvl = 0)
}
\arguments{
\item{object}{an object of class \code{interference}}

\item{allocation1}{the allocation scheme for which to estimate indirect effects}

\item{allocation2}{the allocation scheme for which to estimate indirect effects.
If NULL, then returns all indirect effects compared to allocation1.}

\item{trt.lvl}{Defaults to 0.}
}
\value{
a data.frame with requested values
}
\description{
Retrieves the population average indirect causal effect for
specified allocations:
\eqn{\hat{Y}(0, alpha1) - \hat{Y}(0, alpha2)}{Yhat(0, alpha1) - Yhat(0, alpha2)}.
This is the effect due to the coverage (allocation) levels.

Retreive Indirect Effect estimates
}

