% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/interference.R
\name{interference}
\alias{interference}
\title{Estimate Causal Effects in presence of interference}
\usage{
interference(formula, propensity_integrand,
  loglihood_integrand = propensity_integrand, allocations, data,
  model_method = "glmer", model_options = list(family = binomial(link =
  "logit")), causal_estimation_method = "ipw",
  causal_estimation_options = list(set_NA_to_0 = TRUE, variance_estimation =
  "robust"), conf.level = 0.95, rescale.factor = 1, ...)
}
\arguments{
\item{formula}{}

\item{propensity_integrand}{function used as the integrand in computing the IPW weights.
Defaults to \code{\link{logit_integrand}}.}

\item{loglihood_integrand}{function used as the integrand in computing the scores in
\code{\link{score_calc}}. Defaults to \code{\link{logit_integrand}}.}

\item{allocations}{the allocation (coverage) schemes for which to estimate effects.
Must be in [0, 1].}

\item{data}{the analysis data.frame}

\item{model_method}{One of three options: `glmer`, `glm`, or `oracle`}

\item{model_options}{List of options passed to \code{model_method}}

\item{causal_estimation_method}{currently only supports `ipw`.}

\item{causal_estimation_options}{list of options passed to the corresponding
estimation function. `ipw` is \code{\link{ipw_interference}}}

\item{conf.level}{Confidence level for confidence intervals. Defaults to 0.95.}

\item{rescale.factor}{factor by which to rescale values. Defaults to 1.}

\item{...}{additional arguments passed to other functions such as
\code{\link{glmer}}, \code{\link{grad}}, \code{\link{integrate}}, and \code{integrand} or \code{likelihood}.}

\item{set_NA_to_0}{if TRUE, sets any weights that returned an NA value to 0.
Defaults to TRUE}
}
\value{
Returns a list of overall and group-level IPW point estimates
(the output of \code{\link{ipw_point_estimates}}), overall and group-level IPW
point estimates (using the weight derivatives), scores (the output of
\code{\link{score_matrix_calc}}), the computed weight matrix, the computed
weight derivative array, and a summary.
}
\description{
Estimate Causal Effects in presence of interference
}

