% Generated by roxygen2 (4.0.1): do not edit by hand
\name{ipw_interference}
\alias{ipw_interference}
\title{IPW Interference estimation}
\usage{
ipw_interference(integrand = "logit_integrand",
  likelihood = "logit_integrand", allocations, data, groups, outcome,
  treatment, propensityB = treatment, propensity_formula,
  model_method = "glmer", model_options = list(family = binomial(link =
  "logit")), set_NA_to_0 = TRUE, ...)
}
\arguments{
\item{integrand}{function used as the integrand in computing the IPW weights.
Defaults to \code{\link{logit_integrand}}.}

\item{likelihood}{function used as the integrand in computing the scores in
\code{\link{score_calc}}. Defaults to \code{\link{logit_integrand}}.}

\item{allocations}{the allocation (coverage) schemes for which to estimate effects.
Must be in (0, 1].}

\item{model_method}{One of three options: `glmer`, `glm`, or `oracle`}

\item{model_options}{List of options passed to \code{model_method}}

\item{set_NA_to_0}{if TRUE, sets any weights that returned an NA value to 0.
Defaults to TRUE}

\item{...}{additional arguments passed to other functions such as
\code{\link{glmer}}, \code{\link{grad}}, and \code{integrand} or \code{likelihood}.}

\item{data}{the analysis data.frame}

\item{groups}{quoted name of group variable in \code{data}}

\item{outcome}{quoted name of outcome variable in \code{data}}

\item{treatment}{quoted name of treatment variable in \code{data}}

\item{propensityB}{Optional. If the treatment variable is different for
\code{integrand} and \code{likelihood}, include the quote name of the variable
to be passed to \code{likelihood}. Defaults to \code{treatment}.}

\item{propensity_formula}{formula passed to model_method for estimation of propensity score
parameters.}
}
\value{
Returns a list of overall and group-level IPW point estimates
(the output of \code{\link{ipw_point_estimates}}), overall and group-level IPW
point estimates (using the weight derivatives), scores (the output of
\code{\link{score_matrix_calc}}), the computed weight matrix, the computed
weight derivative array, and a summary.
}
\description{
Prepares the object necessary to compute IPW effect estimates with
\code{\link{calc_effect}}, \code{\link{direct_effect}}, \code{\link{indirect_effect}},
\code{\link{total_effect}}, or \code{\link{overall_effect}}.
}

