% Generated by roxygen2 (4.0.1): do not edit by hand
\name{ipw_point_estimates}
\alias{ipw_point_estimates}
\title{Calculate IPW point estimates}
\usage{
ipw_point_estimates(outcome, groups, treatment, data, weights)
}
\arguments{
\item{outcome}{quoted name of outcome variable in \code{data}}

\item{groups}{quoted name of group variable in \code{data}}

\item{treatment}{quoted name of treatment variable in \code{data}}

\item{weights}{weight matrix/array to use from either \code{\link{wght_matrix}}
or \code{\link{wght_deriv_array}}}
}
\value{
list containing point estimates for marginal outcomes and estimates
per treatment level
}
\description{
Calculate IPW point estimates
}

