% Generated by roxygen2 (4.0.1): do not edit by hand
\name{ipw_point_estimates}
\alias{ipw_point_estimates}
\title{Calculate IPW point estimates}
\usage{
ipw_point_estimates(y, G, A, B, data, weights, rescale.factor,
  set.NA.to.0 = TRUE)
}
\arguments{
\item{y}{vector of unweighted group means (i.e. output from \code{\link{group_means}}}

\item{a}{treatment level (0,1) for which to compute weights. Defaults to NULL which returns marginal.}

\item{weights}{weight matrix/array to use}

\item{rescale.factor}{factor by which to rescale values}

\item{na.rm}{exclude groups missing weights? defaults to FALSE.}
}
\value{
length(alpha) vector of IPW estimates
}
\description{
Calculate IPW point estimates
}

