% Generated by roxygen2 (4.0.1): do not edit by hand
\name{log_likelihood}
\alias{log_likelihood}
\title{Log Likelihood}
\usage{
log_likelihood(x, pos, integrand = logit_integrand, ...)
}
\arguments{
\item{x}{used by \code{\link{grad}} to take the derivative of the
\code{\link{integrate}}(\code{integrand}) with respect to each value of the
\code{param} argument in \code{integrand}}

\item{pos}{}

\item{integrand}{Defaults to logit_integrand}
}
\value{
value of log likelihood
}
\description{
Used by \code{\link{score_matrix_calc}} to compute the log likelihood.
}

