% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/score_functions.R
\name{log_likelihood}
\alias{log_likelihood}
\title{Log Likelihood}
\usage{
log_likelihood(x, pos, integrand = logit_integrand, ...)
}
\arguments{
\item{x}{used by \code{\link{grad}} to take the derivative of the
\code{\link{integrate}}(\code{integrand}) with respect to each value of the
\code{param} argument in \code{integrand}}

\item{pos}{The position of theta for which to take the derivative.}

\item{integrand}{Defaults to logit_integrand}

\item{...}{additional arguments passed to \code{integrand} function.}
}
\value{
value of log likelihood
}
\description{
Used by \code{\link{score_matrix_calc}} to compute the log likelihood.
}

