% Generated by roxygen2 (4.0.1): do not edit by hand
\name{logit_integrand}
\alias{logit_integrand}
\title{Integrand of Pr(A|X)}
\usage{
logit_integrand(b, X, A, fixed.effects, random.effect = NULL, x = NULL,
  pos = NULL, allocation = NULL, randomization = 1,
  integrate.allocation = FALSE)
}
\arguments{
\item{b}{vector argument of values necessary for \code{\link{integrate}}}

\item{x}{Used by \code{\line{grad}} for taking the derivative with respect an element of
params. Only used if \code{pos} is not NULL.}

\item{pos}{The position of theta for which to take the derivative. Defaults to NULL.}

\item{X}{n by length(params) - 1 matrix of covariates. Make sure the order of columns in X corresponds to params.}

\item{params}{p + 1 vector of fixed effects plus the random effect variance. The variance estimate must be the last element.}

\item{A}{vector of observed treatments (0,1)}

\item{allocation}{The allocation strategy. Required if include.alpha == TRUE.
Defaults to NA.}

\item{r}{Randomization probability. Defaults to 1.}

\item{include.allocation}{Either TRUE for including allocation in the product or FALSE
does not include allocation. See \code{\link{wght_calc}} for more information.}
}
\value{
value of the integrand
}
\description{
Computes the following function:
\deqn{\prod_{j=1}^{n} (r h_{j}(b))^{A_j}  (1 - r h_{j}(b))^{1 - A_j}
f_b(b; \theta_b)}{ prod(r * plogis(X * fixef + b)^A *
(1 - r * plogis(X * fixef+ b))^(1 - A)) *
dnorm(sd = sqrt(ranef))}
where \eqn{r} is the randomization scheme. \eqn{X} is the covariate(s) vectors.
\eqn{fixef} is the vector of fixed effects. \eqn{b} is the random (group-level) effect.
\eqn{ranef} is the random effect variance. Used by \code{\link{wght_calc}} and
\code{\link{score_calc}}.
}

