% Generated by roxygen2 (4.0.1): do not edit by hand
\name{overall_effect}
\alias{overall_effect}
\title{Calculate Overall effect estimates}
\usage{
overall_effect(obj, alpha1, alpha2, print = FALSE, rescale.factor = 1,
  conf.level = 0.95)
}
\arguments{
\item{obj}{the name of the object created by \code{\link{run_interference}}}

\item{alpha1}{the allocation scheme for which to estimate overall effects}

\item{alpha2}{the allocation scheme for which to estimate overall effects}

\item{rescale.factor}{factor by which to rescale values. Defaults to 1.}

\item{print}{see \code{\link{calc_effect}}}

\item{conf.level}{see \code{\link{calc_effect}}}
}
\value{
See \code{\link{calc_effect}}.
}
\description{
By default, this function computes:
\eqn{\hat{Y}(alpha1) - \hat{Y}(lpha2)}{Yhat(alpha1) - Yhat(alpha2)}
}

