% Generated by roxygen2 (4.0.1): do not edit by hand
\name{overall_effect}
\alias{overall_effect}
\title{Calculate Overall effect estimates}
\usage{
overall_effect(obj, alpha1, alpha2, print = FALSE, conf.level = 0.95)
}
\arguments{
\item{TBD}{}
}
\value{
TBD
}
\description{
Calculate Overall effect estimates
}

