% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/effect_estimates.R
\name{overall_effect}
\alias{overall_effect}
\title{Calculate Overall Effect Estimates}
\usage{
overall_effect(obj, allocation1, allocation2, print = FALSE,
  rescale.factor = 1, conf.level = 0.95)
}
\arguments{
\item{obj}{the name of the object created by \code{\link{ipw_interference}}}

\item{allocation1}{the allocation scheme for which to estimate overall effects}

\item{allocation2}{the allocation scheme for which to estimate overall effects}

\item{print}{TRUE/FALSE. If TRUE, the point estimates and confidence interval
are printed to the console.}

\item{rescale.factor}{factor by which to rescale values. Defaults to 1.}

\item{conf.level}{Confidence level for confidence intervals. Defaults to 0.95.}
}
\value{
See \code{\link{calc_effect}}.
}
\description{
Computes the population average overall causal effect:
\eqn{\hat{Y}(alpha1) - \hat{Y}(lpha2)}{Yhat(alpha1) - Yhat(alpha2)}
}

