% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utilities.R
\name{overall_effect}
\alias{oe}
\alias{overall_effect}
\title{Retrieve Overall Effect Estimates}
\usage{
overall_effect(object, allocation1, allocation2 = NULL)

oe(object, allocation1, allocation2 = NULL)
}
\arguments{
\item{object}{an object of class \code{interference}}

\item{allocation1}{the allocation scheme for which to estimate overall effects}

\item{allocation2}{the allocation scheme for which to estimate overall effects}
}
\value{
a data.frame with a single row with requested values
}
\description{
Retrieves the population average overall causal effect:
\eqn{\hat{Y}(alpha1) - \hat{Y}(lpha2)}{Yhat(alpha1) - Yhat(alpha2)}
}

