% Generated by roxygen2 (4.0.1): do not edit by hand
\name{run_interference}
\alias{run_interference}
\title{Run Interference}
\usage{
run_interference(f.ab, alphas, data, groups, outcome, A, B, predictors,
  type = "b", rescale.factor = 1)
}
\arguments{
\item{predictors}{}
}
\value{
N X length(theta) matrix of scores
}
\description{
Run Interference
}

