% Generated by roxygen2 (4.0.1): do not edit by hand
\name{run_interference}
\alias{run_interference}
\title{Run Interference}
\usage{
run_interference(f.ab = logit_integrand, likelihood = logit_integrand,
  alphas, data, groups, outcome, treatment, predictors, include.alpha = FALSE,
  propensityB = treatment, family = binomial(link = "logit"),
  known_params = NULL, set.NA.to.0 = TRUE, ...)
}
\arguments{
\item{f.ab}{f.ab function to pass to the argument 'f' of \code{\link{integrate}}.}

\item{alphas}{the allocation schemes for which to estimate effects. Must be in (0, 1].}

\item{data}{the analysis data.frame}

\item{groups}{quoted name of group variable in \code{data}}

\item{outcome}{quoted name of outcome variable in \code{data}}

\item{treatment}{quoted name of treatment variable in \code{data}}

\item{predictors}{character vector of predictor variables in \code{data}}

\item{type}{see \code{\link{wght_calc}}}

\item{propensityB.}{Defaults to \code{treatment}}

\item{family}{passed to \code{\link{glmer}}. See \code{\link{family}}}

\item{known_params}{if the parameter vector is known (e.g. for simulated data),
this argument can be used to pass the known parameters to \code{\link{wght_calc}}
and related functions. If this argument is given, fixed effect parameter estimation
is skipped. Defaults to NULL.}

\item{additional}{arguments passed to other functions. See \code{\link{glmer}}}
}
\value{
Returns a list of ..
}
\description{
Prepares the object necessary to compute IPW effect estimates with
\code{\link{calc_effect}}, \code{\link{direct_effect}}, \code{\link{indirect_effect}},
\code{\link{total_effect}}, or \code{\link{overall_effect}}.
}

