% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/score_functions.R
\name{score_calc}
\alias{score_calc}
\title{Compute scores for a single group}
\usage{
score_calc(integrand = logit_integrand, hide.errors = TRUE, fixed.effects,
  random.effects, ...)
}
\arguments{
\item{integrand}{function to used for the integrand.
Defaults to \code{\link{logit_integrand}}.}

\item{hide.errors}{Hide errors printed from \code{\link{grad}}.
Defaults to true.}

\item{fixed.effects}{vector of fixed effect parameters.}

\item{random.effects}{OPTIONAL vector random effect parameters.}

\item{...}{additional arguments pass to the integrand function.}
}
\value{
length(theta) vector of scores
}
\description{
Used by \code{\link{score_matrix_calc}} to log likelihood derivatives for
a single group.
}

