% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_functions.R
\name{score_matrix}
\alias{score_matrix}
\title{Calculate matrix of log Likelihood derivatives}
\usage{
score_matrix(integrand, X, A, G, parameters, runSilent = TRUE, ...)
}
\arguments{
\item{integrand}{function passed to \code{\link{log_likelihood}}. Defaults to
\code{\link{logit_integrand}}}

\item{X}{covariate matrix}

\item{A}{vector of treatment assignments}

\item{G}{vector of group assignments}

\item{parameters}{vector of parameters passed to \code{integrand}}

\item{runSilent}{If FALSE, prints errors to console. Defaults to TRUE.}

\item{...}{additional arguments passed to \code{integrand} or \code{\link{grad}}.
For example, one can change the \code{method} argument in \code{grad}.}
}
\value{
N X length(params) matrix of scores
}
\description{
Calculate matrix of log Likelihood derivatives
}
