% Generated by roxygen2 (4.0.1): do not edit by hand
\name{summary.interference}
\alias{summary.interference}
\title{Summary of a causal_interference object}
\usage{
summary.interference(object)
}
\arguments{
\item{object}{object of class 'interference'}
}
\description{
Summary of a causal_interference object
}

