% Generated by roxygen2 (4.0.1): do not edit by hand
\name{summary.interference}
\alias{summary.interference}
\title{Prints a summary of an interference object}
\usage{
summary.interference(object, ...)
}
\arguments{
\item{object}{object of class 'interference'}

\item{ignored}{}
}
\description{
Prints a summary of an interference object
}

