% Generated by roxygen2 (4.0.1): do not edit by hand
\name{total_effect}
\alias{total_effect}
\title{Calculate indirect effect estimates}
\usage{
total_effect(obj, alpha1, alpha2, trt.lvl1 = "0", trt.lvl2 = "1")
}
\arguments{
\item{TBD}{}
}
\value{
TBD
}
\description{
Calculate indirect effect estimates
}

