% Generated by roxygen2 (4.0.1): do not edit by hand
\name{total_effect}
\alias{total_effect}
\title{Calculate Total effect estimates}
\usage{
total_effect(obj, alpha1, alpha2, trt.lvl1 = "0", trt.lvl2 = "1",
  print = FALSE)
}
\arguments{
\item{TBD}{}
}
\value{
TBD
}
\description{
Calculate Total effect estimates
}

