% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/utilities.R
\name{total_effect}
\alias{te}
\alias{total_effect}
\title{Retrieve Total Effect estimates}
\usage{
total_effect(object, allocation1, allocation2 = NULL, trt.lvl1 = 0)

te(object, allocation1, allocation2 = NULL, trt.lvl1 = 0)
}
\arguments{
\item{object}{an object of class \code{interference}}

\item{allocation1}{the allocation scheme for which to estimate total effects}

\item{allocation2}{the allocation scheme for which to estimate total effects
If NULL, then returns all indirect effects compared to allocation1.}

\item{trt.lvl1}{Defaults to 0.}
}
\value{
a data.frame with requested values
}
\description{
Retrieves the population average total causal effect for
specified allocations:
\eqn{\hat{Y}(0, alpha1) - \hat{Y}(1, alpha2)}{Yhat(0, alpha1) - Yhat(1, alpha2)}
}

