% Generated by roxygen2 (4.0.1): do not edit by hand
\name{V_matrix}
\alias{V_matrix}
\title{V Matrix}
\usage{
V_matrix(Bscores, ipw_obj, alpha1, trt.lvl1, alpha2 = NA, trt.lvl2 = NA,
  effect, marginal, set.NA.to.0 = TRUE)
}
\arguments{
\item{predictors}{}
}
\value{
V matrix
}
\description{
V Matrix
}

