% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/datasets.R
\name{voters}
\alias{voters}
\title{Voting Contagion Experiment Data}
\format{a dataset with 21 variables and 7722 rows
\describe{
   \item{family}{household ID}
   \item{denver}{1 = subject in Denver, 0 = Minneapolis}
   \item{treatment}{1 = voting encouragement, 2 = recycling message, 3 = not contacted}
   \item{reached}{1 = subject answered door, 0 = not}
   \item{hsecontact}{1 = household contacted by canvassers, 0 = not}
   \item{voted02p}{1 = voted in '02 primary, 0 = not}
   \item{party}{party affiliation}
   \item{age}{age}
   \item{gender}{gender}
 }}
\source{
\url{http://www3.nd.edu/~dnickers/data.php}
}
\description{
A dataset of a voting contagion experiment. For complete documentation, see
the 'Minneapolis and Denver Contagion Experiment' files on \href{http://www3.nd.edu/~dnickers/data.php}{David Nickerson's
website.} See also \href{http://dx.doi.org/10.1017/S0003055408080039}{Nickerson 2008} for more
details.
}
\details{
Just the variables used in the package vignette are documented here.
}

