% Generated by roxygen2 (4.0.1): do not edit by hand
\name{wght_calc}
\alias{wght_calc}
\title{Compute IPW weight}
\usage{
wght_calc(integrand, allocation, x = NULL, pos = NULL, ...)
}
\arguments{
\item{integrand}{function to pass to the argument 'f' of \code{\link{integrate}}.}

\item{type}{see description. Defaults to 'b'.}

\item{x}{necessary argument for \code{\link{grad}}. Defaults to NA, so if
not evaluting a derivative with \code{\link{wght_deriv_calc}}, this can be
ignored.}

\item{pos}{necessary taking a derivative when using to \code{\link{PrAX_integrand}}
Defaults to NA, so if not evaluting a derivative with \code{\link{wght_deriv_calc}},
this can be ignored.}

\item{...}{other arguments passed to integrand.}
}
\value{
scalar result of the integral
}
\description{
Calculates the IPW for a single group. Used by \code{\link{wght_matrix}} to
create a matrix of weights for each group and allocation scheme.
}
\details{
Type c =
\deqn{\frac{\prod_{j=1}^n \alpha^A_j (1 - \alpha)^(1-
A_j)}{Pr(A|X)}}{prod(alpha^A(1 - alpha)^A) / integrate(PrAX_integrand)}
Type b =
\deqn{\frac{1}{Pr(A|X)}}{1 / integrate(PrAX_integrand)}

Type b incorporates the numerator of type c into the \eqn{Pr(A|X)} integral,
resulting in a slower computation but more accurate results (especially
for large groups)
}

