% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/wght_calc.R
\name{wght_calc}
\alias{wght_calc}
\title{Compute IPW weight}
\usage{
wght_calc(integrand, allocation, x = NULL, pos = NULL, ...)
}
\arguments{
\item{integrand}{function to pass to the argument 'f' of \code{\link{integrate}}.}

\item{allocation}{the allocation ratio for which to compute the weight}

\item{x}{OPTIONAL argument necessary for \code{\link{grad}} when using \code{\link{wght_deriv_calc}}.}

\item{pos}{OPTIONAL argument necessary for \code{\link{logit_integrand}} when using \code{\link{wght_deriv_calc}}.}

\item{...}{other arguments passed to integrand.}
}
\value{
scalar result of the integral
}
\description{
Calculates the IPW for a single group. Used by \code{\link{wght_matrix}} to
create a matrix of weights for each group and allocation scheme.
}
\details{
If \code{integrate.allocation} is an argument in the integrand function and
\code{integrate.allocation == TRUE}, then the weight is calcuated as:

\deqn{\frac{1}{Pr(A|X)}}{1 / integrate(integrand)}

Otherwise, the weight is computed by:
\deqn{\frac{\prod_{j=1}^n \alpha^A_j (1 - \alpha)^(1-
A_j)}{Pr(A|X)}}{prod(allocation^A(1 - allocation)^A) / integrate(integrand)}
}

