% Generated by roxygen2 (4.0.1): do not edit by hand
\name{wght_calc}
\alias{wght_calc}
\title{Compute a group weight}
\usage{
wght_calc(type, A, alpha, ...)
}
\arguments{
\item{f}{the function to integrad. Defaults to \code{\link{PrAX_integrand}}}

\item{llimit}{lower limit of the integral defaults to \code{-Inf}}

\item{ulimit}{upper limit of the integral defaults to \code{Inf}}

\item{type}{see description}

\item{alpha}{allocation strategy}

\item{A}{vector of treatments}

\item{...}{other arguments passed to integrand}
}
\value{
scalar which is the result of the integral
}
\description{
Type c =
\eqn{\frac{\pi}{Pr(A|X)}}
Type b =
\eqn{\frac{1}{Pr(A|X)}}
where the \eqn{\pi} term is brought into the dominator
}

