% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wght_deriv_array.R
\name{wght_deriv_array}
\alias{wght_deriv_array}
\title{Create an array of group weight derivatives}
\usage{
wght_deriv_array(parameters, integrand, allocations, X, A, G,
  runSilent = TRUE, ...)
}
\arguments{
\item{parameters}{vector of parameters passed to \code{integrand}}

\item{integrand}{the function to passed to the argument 'f' of \code{\link{integrate}},
which is part of \code{\link{wght_calc}}.}

\item{allocations}{coverage levels in (0, 1), possibly (probably) vector valued}

\item{X}{covariate matrix}

\item{A}{vector of treatment assignments}

\item{G}{vector of group assignments}

\item{runSilent}{if FALSE, errors are printed to console. Defaults to TRUE.}

\item{...}{additional arguments passed to integrand}
}
\value{
a length(unique(group)) X length(params) X length(alphas) array of
group weight derivatives
}
\description{
Uses \code{\link{wght_deriv_calc}} to compute the weight derivatives for each
group per coverage level
}

