% Generated by roxygen2 (4.0.1): do not edit by hand
\name{wght_deriv_array}
\alias{wght_deriv_array}
\title{Create an array of group weight derivative}
\usage{
wght_deriv_array(f.ab, alphas, data, groups, predictors, treatment, params,
  include.alpha, ...)
}
\arguments{
\item{alphas}{coverage levels in (0, 1), possibly (probably) vector valued}

\item{data}{data frame}

\item{groups}{quoted string for name of variable in data containing group membership}

\item{predictors}{character vector of names of predictor variables in data}

\item{A}{character vector of name of treatment variable in data}

\item{params}{p + 1 vector of fixed effects plus the random effect variance. The variance estimate must be last.}

\item{type}{type of weight to compute. See \code{\link{wght_calc}}}

\item{...}{additional arguments passed to f.ab}
}
\value{
a length(unique(group)) X length(alphas) matrix of group weights
}
\description{
- creates a length(group) X length(theta) X length(alphas) array
}

