% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/wght_deriv_calc.R
\name{wght_deriv_calc}
\alias{wght_deriv_calc}
\title{Compute the derivative(s) of a weight}
\usage{
wght_deriv_calc(integrand, fixed.effects, random.effects = NULL, allocation,
  hide.errors = TRUE, ...)
}
\arguments{
\item{integrand}{the function to passed to the argument 'f' of \code{\link{integrate}},
which is part of \code{\link{wght_calc}}.}

\item{fixed.effects}{vector of fixed effect parameters}

\item{random.effects}{vector random effect parameters}

\item{allocation}{the allocation ratio for which to compute the weights}

\item{hide.errors}{print \code{grad} error messages. Defaults to TRUE.}

\item{...}{additional arguments passed to integrand}
}
\value{
vector of derivatives with respect to element of params
}
\description{
Takes the derivative of the \code{\link{wght_calc}} function with respect to each
parameter in \code{params}.
}

