% Generated by roxygen2 (4.0.1): do not edit by hand
\name{wght_deriv_calc}
\alias{wght_deriv_calc}
\title{Compute the derivative(s) of a weight}
\usage{
wght_deriv_calc(integrand, params, allocation, hide.errors = TRUE, ...)
}
\arguments{
\item{integrand}{the function to passed to the argument 'f' of \code{\link{integrate}},
which is part of \code{\link{wght_calc}}.}

\item{params}{parameters with which to take derivatives with respect to}

\item{allocation}{the allocation ratio for which to compute the weights}

\item{hide.errors}{print \code{grad} error messages. Defaults to TRUE.}

\item{...}{additional arguments passed to integrand}
}
\value{
vector of derivatives with respect to element of params
}
\description{
Takes the derivative of the \code{\link{wght_calc}} function with respect to each
parameter in \code{params}.
}

