% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wght_functions.R
\name{wght_deriv_calc}
\alias{wght_deriv_calc}
\title{Compute the derivative(s) of a weight}
\usage{
wght_deriv_calc(integrand, parameters, allocation,
  integrate_allocation = TRUE, ...)
}
\arguments{
\item{integrand}{function to pass to the argument 'f' of \code{\link{integrate}}.}

\item{parameters}{vector of parameter values}

\item{allocation}{the allocation ratio for which to compute the weight}

\item{integrate_allocation}{Indicator of whether the integrand function uses 
the allocation parameter. Defaults to TRUE.}

\item{...}{other arguments passed to integrand.}
}
\value{
vector of derivatives with respect to element of params
}
\description{
Takes the derivative of the \code{\link{wght_calc}} function with respect to each
parameter in \code{params}.
}
