% Generated by roxygen2 (4.0.1): do not edit by hand
\name{wght_matrix}
\alias{wght_matrix}
\title{Create a matrix of group IP weights}
\usage{
wght_matrix(integrand = logit_integrand, allocations, X, A, G, fixed.effects,
  random.effects = NULL, ...)
}
\arguments{
\item{integrand}{the function used in the weight calculation. Defaults to
\code{\link{logit_integrand}}}

\item{allocations}{coverage levels in [0, 1]. Can be vector.}

\item{X}{covariate matrix}

\item{A}{vector of treatment assignments}

\item{G}{vector of group assignments}

\item{fixed.effects}{vector of fixed effect parameters}

\item{random.effects}{OPTIONAL vector random effect parameters}

\item{...}{additional arguments passed to \code{integrand}}
}
\value{
a length(unique(group)) X length(alphas) matrix of group weights
}
\description{
Creates a number of groups by number of allocation schemes matrix of group weights.
Allocation schemes are selected by the user.
}
\details{
Groups should be numbered 1, ..., N
}

