% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wght_matrix.R
\name{wght_matrix}
\alias{wght_matrix}
\title{Creates a number of groups by number of allocation schemes matrix of group weights.
Allocation schemes are selected by the user.}
\usage{
wght_matrix(integrand, allocations, X, A, G, parameters, runSilent = FALSE,
  ...)
}
\arguments{
\item{integrand}{the function used in the weight calculation. Defaults to
\code{\link{logit_integrand}}}

\item{allocations}{coverage levels in [0, 1]. Can be vector.}

\item{X}{covariate matrix}

\item{A}{vector of treatment assignments}

\item{G}{vector of group assignments}

\item{...}{additional arguments passed to \code{integrand}}

\item{fixed.effects}{vector of fixed effect parameters}

\item{random.effects}{OPTIONAL vector random effect parameters}
}
\value{
a length(unique(group)) X length(alphas) matrix of group weights
}
\description{
Groups should be numbered 1, ..., N
}

