% Generated by roxygen2 (4.0.1): do not edit by hand
\name{wght_matrix}
\alias{wght_matrix}
\title{Create a matrix of group weights}
\usage{
wght_matrix(integrand = logit_integrand, allocations, data, groups,
  predictors, treatment, params, ...)
}
\arguments{
\item{integrand}{the function used in the weight calculation. Defaults to
\code{\link{logit_integrand}}}

\item{allocations}{coverage levels in (0, 1), possibly (probably) vector valued}

\item{data}{data frame}

\item{groups}{quoted string for name of variable in data containing group membership}

\item{predictors}{character vector of names of predictor variables in data}

\item{A}{character vector of name of treatment variable in data}

\item{params}{p + 1 vector of fixed effects plus the random effect variance.
The variance estimate must be last.}

\item{...}{additional arguments passed to \code{integrand}}
}
\value{
a length(unique(group)) X length(alphas) matrix of group weights
}
\description{
Creates a number of groups by number of allocation schemes matrix of group weights.
Allocation schemes are selected by the user.
}
\details{
Groups should be numbered 1, ..., N

\code{predictors} does not need to have a value corresponding to reference parameter.
The function adds a column of 1s with \code{cbind()} as a first step.
}

